requirements.txt
nic-api
dnspython>=2.0.0
tld
cryptography
cx_Freeze
pytest